VE_TABLE/requirements.txt
pyqt5
pyqtgraph
numpy
pandas
OBD
pyserial>=3.5
bleak

